pyproject.toml
[project]
name = "medical-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bert-score>=0.3.13",
    "datasets>=4.0.0",
    "evaluate>=0.4.5",
    "huggingface-hub[cli]>=0.34.4",
    "ipykernel>=6.30.1",
    "peft>=0.17.0",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "transformers[torch]>=4.55.2",
    "trl>=0.21.0",
]
